+++ CHANGELOG.md
@@ -15,19 +15,22 @@
 
 ### Changed
 
+- allow multiple embeddings in conclass (July 2020)
 - Improved `plotDEheatmap` function
 - Fixed bug with `balancing.factor.per.sample` in `buildGraph`
 - Fixed some installation problems
 - Improve R6 documentation
+- Changed `std::cout` to `Rcpp::Rcout` (July 2020)
+- Revised README, vignettes (July 2020)
+
 
 ### Added
 
-- Changed `std::cout` to `Rcpp::Rcout` (July 2020)
-- Revised README, vignettes (July 2020)
+- multiple embeddings in Conos object (July 2020)
+- Write to HDF5 for `saveConosForScanPy()` (July 2020)
 - added `ht_opt$message = FALSE` for ComplexHeatmap (July 2020)
 - Added checks for `getPerCellTypeDE()` for errors, removing NAs (July 2020)
 - Added `ht_opt$message = FALSE` for ComplexHeatmap (July 2020)
-- Write to HDF5 for `saveConosForScanPy()` (July 2020)
 - LICENSE (July 2020)
 
 ### Removed
@@ -35,6 +38,7 @@
 - Removed `getCorrectionVector()` and `getPerCellTypeDECorrected` (2 July 2020)
 - Removed all neighborhood averaging via `neighborhood.average` (4 July 2020)
 - Removed `raster.height` and `raster.width` from `con$plotEmbedding()`, given these parameters are defunct with rewrite of `ggrastr` (v0.2.0)[https://github.com/VPetukhov/ggrastr/releases/tag/v0.2.0]
+
 
 ## [1.3.0] - 2020-19-3